+++ Model/Management/PlaceOrder.php
@@ -23,6 +23,7 @@ use Qliro\QliroOne\Model\Logger\Manager as LogManager;
 use Qliro\QliroOne\Model\Order\OrderPlacer;
 use Qliro\QliroOne\Model\Order\OrganizationNumber;
 use Qliro\QliroOne\Model\QliroOrder\Converter\QuoteFromOrderConverter;
+use Qliro\QliroOne\Model\QliroOrder\RoundingAdjustmentStamp;
 use Qliro\QliroOne\Model\ResourceModel\Lock;
 use Qliro\QliroOne\Model\Exception\TerminalException;
 use Qliro\QliroOne\Model\Exception\FailToLockException;
@@ -114,6 +115,11 @@ class PlaceOrder extends AbstractManagement
     private $organizationNumber;
 
     /**
+     * @var \Qliro\QliroOne\Model\QliroOrder\RoundingAdjustmentStamp|null
+     */
+    private ?RoundingAdjustmentStamp $roundingAdjustmentStamp;
+
+    /**
      * Inject dependencies
      *
      * @param Config $qliroConfig
@@ -132,6 +138,7 @@ class PlaceOrder extends AbstractManagement
      * @param Payment $paymentManagement
      * @param RecurringDataService $recurringDataService
      * @param OrganizationNumber $organizationNumber
+     * @param RoundingAdjustmentStamp|null $roundingAdjustmentStamp
      */
     public function __construct(
         Config $qliroConfig,
@@ -149,7 +156,8 @@ class PlaceOrder extends AbstractManagement
         Quote $quoteManagement,
         Payment $paymentManagement,
         RecurringDataService $recurringDataService,
-        OrganizationNumber $organizationNumber
+        OrganizationNumber $organizationNumber,
+        ?RoundingAdjustmentStamp $roundingAdjustmentStamp = null
     ) {
         $this->qliroConfig = $qliroConfig;
         $this->merchantApi = $merchantApi;
@@ -167,6 +175,8 @@ class PlaceOrder extends AbstractManagement
         $this->paymentManagement = $paymentManagement;
         $this->recurringDataService = $recurringDataService;
         $this->organizationNumber = $organizationNumber;
+        // Optional so a store constructing this class with the old signature keeps working
+        $this->roundingAdjustmentStamp = $roundingAdjustmentStamp;
     }
 
     /**
@@ -349,7 +359,11 @@ class PlaceOrder extends AbstractManagement
                     );
 
                     $this->quoteFromOrderConverter->convert($qliroOrder, $this->getQuote());
-                    $this->addAdditionalInfoToQuote($link, $qliroOrder->getPaymentMethod());
+                    $this->addAdditionalInfoToQuote(
+                        $link,
+                        $qliroOrder->getPaymentMethod(),
+                        $qliroOrder->getOrderItems() ?? []
+                    );
                     $this->addAdditionalShippingInfoToQuote($qliroOrder);
                     $this->quoteManagement->setQuote($this->getQuote())->recalculateAndSaveQuote();
 
@@ -582,9 +596,10 @@ class PlaceOrder extends AbstractManagement
      *
      * @param \Qliro\QliroOne\Api\Data\LinkInterface $link
      * @param \Qliro\QliroOne\Api\Data\QliroOrderPaymentMethodInterface $paymentMethod
+     * @param \Qliro\QliroOne\Api\Data\QliroOrderItemInterface[] $reservationItems
      * @throws \Magento\Framework\Exception\LocalizedException
      */
-    private function addAdditionalInfoToQuote($link, $paymentMethod)
+    private function addAdditionalInfoToQuote($link, $paymentMethod, array $reservationItems = [])
     {
         $payment = $this->getQuote()->getPayment();
         $payment->setAdditionalInformation(Config::QLIROONE_ADDITIONAL_INFO_QLIRO_ORDER_ID, $link->getQliroOrderId());
@@ -597,6 +612,7 @@ class PlaceOrder extends AbstractManagement
             Config::QLIROONE_ADDITIONAL_INFO_LINE_REFERENCE_CARRIES_ITEM_ID,
             true
         );
+        $this->roundingAdjustmentStamp?->record($this->getQuote(), $reservationItems);
 
         if ($paymentMethod) {
             $payment->setAdditionalInformation(